Genomic scores implementations moved to new module.
