+++ dae/dae/genomic_resources/tests/test_cli_repair.py
@@ -189,257 +189,257 @@ def test_repo_repair_dry_run(proto_fixture):
 #     ]
 
 
-def test_resource_repair_no_update_message(
-        proto_fixture, capsys, caplog):
-    # Given
-    path, _proto = proto_fixture
-    cli_manage([
-        "resource-repair", "-R", str(path), "-r", "one", "-j", "1"])
-    _, _ = capsys.readouterr()
+# def test_resource_repair_no_update_message(
+#         proto_fixture, capsys, caplog):
+#     # Given
+#     path, _proto = proto_fixture
+#     cli_manage([
+#         "resource-repair", "-R", str(path), "-r", "one", "-j", "1"])
+#     _, _ = capsys.readouterr()
 
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "resource-repair", "--dry-run",
-            "-R", str(path), "-r", "one", "-j", "1"])
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "resource-repair", "--dry-run",
+#             "-R", str(path), "-r", "one", "-j", "1"])
 
-    # Then
-    out, err = capsys.readouterr()
-    assert out == ""
-    assert err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> is up to date"),
-        ("grr_manage", logging.INFO,
-         "<one> statistics hash is up to date"),
-    ]
-
-
-def test_repo_repair_no_update_message(
-        proto_fixture, capsys, caplog):
-    # Given
-    path, _proto = proto_fixture
-    cli_manage([
-        "repo-repair", "-R", str(path), "-j", "1"])
-    _, _ = capsys.readouterr()
-
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
-
-    # Then
-    out, err = capsys.readouterr()
-    assert out == ""
-    assert err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> is up to date"),
-        ("grr_manage", logging.INFO,
-         "<one> statistics hash is up to date"),
-    ]
+#     # Then
+#     out, err = capsys.readouterr()
+#     assert out == ""
+#     assert err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> is up to date"),
+#         ("grr_manage", logging.INFO,
+#          "<one> statistics hash is up to date"),
+#     ]
 
 
-def test_resource_repair_dry_run_needs_manifest_update_message(
-        proto_fixture, capsys, caplog):
-    # Given
-    path, proto = proto_fixture
-    cli_manage([
-        "resource-repair", "-R", str(path), "-r", "one", "-j", "1"])
+# def test_repo_repair_no_update_message(
+#         proto_fixture, capsys, caplog):
+#     # Given
+#     path, _proto = proto_fixture
+#     cli_manage([
+#         "repo-repair", "-R", str(path), "-j", "1"])
+#     _, _ = capsys.readouterr()
 
-    res = proto.get_resource("one")
-    with res.open_raw_file("data.txt", "wt") as outfile:
-        outfile.write("alabala2")
-    _, _ = capsys.readouterr()
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
 
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "resource-repair", "--dry-run",
-            "-R", str(path), "-r", "one",
-            "-j", "1"])
-
-    # Then
-    captured = capsys.readouterr()
-    assert captured.out == ""
-    assert captured.err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> should be updated; entries to update in manifest "
-         "['data.txt']"),
-        ("grr_manage", logging.INFO,
-         "Manifest of <one> needs update, cannot check statistics"),
-    ]
+#     # Then
+#     out, err = capsys.readouterr()
+#     assert out == ""
+#     assert err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> is up to date"),
+#         ("grr_manage", logging.INFO,
+#          "<one> statistics hash is up to date"),
+#     ]
 
 
-def test_repo_repair_dry_run_needs_manifest_update_message(
-        proto_fixture, capsys, caplog):
-    # Given
-    path, proto = proto_fixture
-    cli_manage([
-        "repo-repair", "-R", str(path), "-j", "1"])
-    res = proto.get_resource("one")
-    with res.open_raw_file("data.txt", "wt") as outfile:
-        outfile.write("alabala2")
-    _, _ = capsys.readouterr()
+# def test_resource_repair_dry_run_needs_manifest_update_message(
+#         proto_fixture, capsys, caplog):
+#     # Given
+#     path, proto = proto_fixture
+#     cli_manage([
+#         "resource-repair", "-R", str(path), "-r", "one", "-j", "1"])
 
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
+#     res = proto.get_resource("one")
+#     with res.open_raw_file("data.txt", "wt") as outfile:
+#         outfile.write("alabala2")
+#     _, _ = capsys.readouterr()
 
-    # Then
-    captured = capsys.readouterr()
-    assert captured.out == ""
-    assert captured.err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> should be updated; entries to update in manifest "
-         "['data.txt']"),
-        ("grr_manage", logging.INFO,
-         "Manifest of <one> needs update, cannot check statistics"),
-    ]
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "resource-repair", "--dry-run",
+#             "-R", str(path), "-r", "one",
+#             "-j", "1"])
 
-
-def test_resource_repair_dry_run_needs_manifest_and_histogram_update_message(
-        proto_fixture, capsys, caplog):
-    # Given
-    path, _proto = proto_fixture
-    cli_manage([
-        "resource-repair", "-R", str(path), "-r", "one", "-j", "1"])
-
-    setup_tabix(
-        path / "one" / "data.txt.gz",
-        """
-        #chrom  pos_begin  pos_end  s1    s2
-        1       10         15       0.02  1.02
-        1       17         19       0.03  1.03
-        1       22         25       0.04  1.04
-        2       5          80       0.01  2.01
-        3       10         11       0.02  2.02
-        """, seq_col=0, start_col=1, end_col=2, line_skip=1, force=True)
-    _, _ = capsys.readouterr()
-
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "resource-repair", "--dry-run",
-            "-R", str(path), "-r", "one",
-            "-j", "1"])
-
-    # Then
-    captured = capsys.readouterr()
-    assert captured.out == ""
-    assert captured.err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> should be updated; entries to update in manifest "
-         "['data.txt.gz', 'data.txt.gz.tbi']"),
-        ("grr_manage", logging.INFO,
-         "Manifest of <one> needs update, cannot check statistics"),
-    ]
-
-    # And after that::
-    # Given
-    cli_manage([
-        "resource-manifest", "-R", str(path), "-r", "one", ])
-    _, _ = capsys.readouterr()
-    caplog.clear()
-
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "resource-repair", "--dry-run",
-            "-R", str(path), "-r", "one",
-            "-j", "1"])
-
-    # Then
-    captured = capsys.readouterr()
-    assert captured.out == ""
-    assert captured.err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> is up to date"),
-        ("grr_manage", logging.INFO,
-         "Stored hash for <one> is outdated, need update"),
-        ("grr_manage", logging.INFO,
-         "Statistics of <one> need update"),
-    ]
+#     # Then
+#     captured = capsys.readouterr()
+#     assert captured.out == ""
+#     assert captured.err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> should be updated; entries to update in manifest "
+#          "['data.txt']"),
+#         ("grr_manage", logging.INFO,
+#          "Manifest of <one> needs update, cannot check statistics"),
+#     ]
 
 
-def test_repo_repair_dry_run_needs_manifest_and_histogram_update_message(
-        proto_fixture, capsys, caplog):
-    # Given
-    path, _proto = proto_fixture
-    cli_manage([
-        "repo-repair", "-R", str(path), "-j", "1"])
+# def test_repo_repair_dry_run_needs_manifest_update_message(
+#         proto_fixture, capsys, caplog):
+#     # Given
+#     path, proto = proto_fixture
+#     cli_manage([
+#         "repo-repair", "-R", str(path), "-j", "1"])
+#     res = proto.get_resource("one")
+#     with res.open_raw_file("data.txt", "wt") as outfile:
+#         outfile.write("alabala2")
+#     _, _ = capsys.readouterr()
 
-    setup_tabix(
-        path / "one" / "data.txt.gz",
-        """
-        #chrom  pos_begin  pos_end  s1    s2
-        1       10         15       0.02  1.02
-        1       17         19       0.03  1.03
-        1       22         25       0.04  1.04
-        2       5          80       0.01  2.01
-        3       10         11       0.02  2.02
-        """, seq_col=0, start_col=1, end_col=2, line_skip=1, force=True)
-    _, _ = capsys.readouterr()
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
 
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
+#     # Then
+#     captured = capsys.readouterr()
+#     assert captured.out == ""
+#     assert captured.err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> should be updated; entries to update in manifest "
+#          "['data.txt']"),
+#         ("grr_manage", logging.INFO,
+#          "Manifest of <one> needs update, cannot check statistics"),
+#     ]
 
-    # Then
-    captured = capsys.readouterr()
-    assert captured.out == ""
-    assert captured.err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> should be updated; entries to update in manifest "
-         "['data.txt.gz', 'data.txt.gz.tbi']"),
-        ("grr_manage", logging.INFO,
-         "Manifest of <one> needs update, cannot check statistics"),
-    ]
 
-    # And after that::
-    # Given
-    cli_manage([
-        "repo-manifest", "-R", str(path), ])
-    _, _ = capsys.readouterr()
-    caplog.clear()
+# def test_resource_repair_dry_run_needs_manifest_and_histogram_update_message(
+#         proto_fixture, capsys, caplog):
+#     # Given
+#     path, _proto = proto_fixture
+#     cli_manage([
+#         "resource-repair", "-R", str(path), "-r", "one", "-j", "1"])
 
-    # When
-    with caplog.at_level(logging.INFO):
-        cli_manage([
-            "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
+#     setup_tabix(
+#         path / "one" / "data.txt.gz",
+#         """
+#         #chrom  pos_begin  pos_end  s1    s2
+#         1       10         15       0.02  1.02
+#         1       17         19       0.03  1.03
+#         1       22         25       0.04  1.04
+#         2       5          80       0.01  2.01
+#         3       10         11       0.02  2.02
+#         """, seq_col=0, start_col=1, end_col=2, line_skip=1, force=True)
+#     _, _ = capsys.readouterr()
 
-    # Then
-    captured = capsys.readouterr()
-    assert captured.out == ""
-    assert captured.err == ""
-    assert caplog.record_tuples == [
-        ("dae.genomic_resources.repository_factory", logging.INFO,
-         "using default GRR definitions"),
-        ("grr_manage", logging.INFO,
-         "manifest of <one> is up to date"),
-        ("grr_manage", logging.INFO,
-         "Stored hash for <one> is outdated, need update"),
-        ("grr_manage", logging.INFO,
-         "Statistics of <one> need update"),
-    ]
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "resource-repair", "--dry-run",
+#             "-R", str(path), "-r", "one",
+#             "-j", "1"])
+
+#     # Then
+#     captured = capsys.readouterr()
+#     assert captured.out == ""
+#     assert captured.err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> should be updated; entries to update in manifest "
+#          "['data.txt.gz', 'data.txt.gz.tbi']"),
+#         ("grr_manage", logging.INFO,
+#          "Manifest of <one> needs update, cannot check statistics"),
+#     ]
+
+#     # And after that::
+#     # Given
+#     cli_manage([
+#         "resource-manifest", "-R", str(path), "-r", "one", ])
+#     _, _ = capsys.readouterr()
+#     caplog.clear()
+
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "resource-repair", "--dry-run",
+#             "-R", str(path), "-r", "one",
+#             "-j", "1"])
+
+#     # Then
+#     captured = capsys.readouterr()
+#     assert captured.out == ""
+#     assert captured.err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> is up to date"),
+#         ("grr_manage", logging.INFO,
+#          "Stored hash for <one> is outdated, need update"),
+#         ("grr_manage", logging.INFO,
+#          "Statistics of <one> need update"),
+#     ]
+
+
+# def test_repo_repair_dry_run_needs_manifest_and_histogram_update_message(
+#         proto_fixture, capsys, caplog):
+#     # Given
+#     path, _proto = proto_fixture
+#     cli_manage([
+#         "repo-repair", "-R", str(path), "-j", "1"])
+
+#     setup_tabix(
+#         path / "one" / "data.txt.gz",
+#         """
+#         #chrom  pos_begin  pos_end  s1    s2
+#         1       10         15       0.02  1.02
+#         1       17         19       0.03  1.03
+#         1       22         25       0.04  1.04
+#         2       5          80       0.01  2.01
+#         3       10         11       0.02  2.02
+#         """, seq_col=0, start_col=1, end_col=2, line_skip=1, force=True)
+#     _, _ = capsys.readouterr()
+
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
+
+#     # Then
+#     captured = capsys.readouterr()
+#     assert captured.out == ""
+#     assert captured.err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> should be updated; entries to update in manifest "
+#          "['data.txt.gz', 'data.txt.gz.tbi']"),
+#         ("grr_manage", logging.INFO,
+#          "Manifest of <one> needs update, cannot check statistics"),
+#     ]
+
+#     # And after that::
+#     # Given
+#     cli_manage([
+#         "repo-manifest", "-R", str(path), ])
+#     _, _ = capsys.readouterr()
+#     caplog.clear()
+
+#     # When
+#     with caplog.at_level(logging.INFO):
+#         cli_manage([
+#             "repo-repair", "--dry-run", "-R", str(path), "-j", "1"])
+
+#     # Then
+#     captured = capsys.readouterr()
+#     assert captured.out == ""
+#     assert captured.err == ""
+#     assert caplog.record_tuples == [
+#         ("dae.genomic_resources.repository_factory", logging.INFO,
+#          "using default GRR definitions"),
+#         ("grr_manage", logging.INFO,
+#          "manifest of <one> is up to date"),
+#         ("grr_manage", logging.INFO,
+#          "Stored hash for <one> is outdated, need update"),
+#         ("grr_manage", logging.INFO,
+#          "Statistics of <one> need update"),
+#     ]
