+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/CADD/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=cadd_raw
     width="200"></td>
 
-<td>(-3.669, 15.616)</td>
+<td>(nan, nan)</td>
 
 </tr><tr class="score-definition">
 
@@ -94,7 +94,7 @@ h3,h4 {
     title=cadd_phred
     width="200"></td>
 
-<td>(0.001, 52.000)</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -133,6 +133,55 @@ h3,h4 {
         </td>
         <td class="nowrap">606.0 B</td>
         <td class="nowrap">77547fdd631548808bd6f733e39b9a37</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_phred.png'>statistics/histogram_cadd_phred.png</a>
+        </td>
+        <td class="nowrap">18.54 KB</td>
+        <td class="nowrap">79ab6b898617b641b4a86dc1f138fbb5</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_phred.yaml'>statistics/histogram_cadd_phred.yaml</a>
+        </td>
+        <td class="nowrap">2.38 KB</td>
+        <td class="nowrap">ea436c3bb52cc7302212a1f167570a25</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_raw.png'>statistics/histogram_cadd_raw.png</a>
+        </td>
+        <td class="nowrap">19.08 KB</td>
+        <td class="nowrap">e1af790e58efc438f4621220b2b4a033</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_raw.yaml'>statistics/histogram_cadd_raw.yaml</a>
+        </td>
+        <td class="nowrap">2.57 KB</td>
+        <td class="nowrap">bdbc630a5fc635026f9a3836223df4d0</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_cadd_phred.yaml'>statistics/min_max_cadd_phred.yaml</a>
+        </td>
+        <td class="nowrap">42.0 B</td>
+        <td class="nowrap">79346b3fc62b019545b8d9299eb31dc3</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_cadd_raw.yaml'>statistics/min_max_cadd_raw.yaml</a>
+        </td>
+        <td class="nowrap">49.0 B</td>
+        <td class="nowrap">6e9e93dcb90385720a07807887b34f15</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">9.97 KB</td>
+        <td class="nowrap">56813a1be3862f38ba21f37b896b6f8b</td>
     </tr>
 </tbody>
 </table>
