+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/GATK_ResourceBundle_5777_b37_phiX174_short/gene_models/refGene_201309/index.html
@@ -98,6 +98,20 @@ Old gene models used by GPF up to 2019-02.
         <td class="nowrap">3.67 KB</td>
         <td class="nowrap">a42701941f92c7b999a4f8f26e4a3a47</td>
     </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats.yaml'>statistics/stats.yaml</a>
+        </td>
+        <td class="nowrap">198.0 B</td>
+        <td class="nowrap">90afd7bdd4ca41745aafa92369763180</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">127.0 B</td>
+        <td class="nowrap">9e80b34285fd49637b6ccf862fbcbf38</td>
+    </tr>
 </tbody>
 </table>
