+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/GATK_ResourceBundle_5777_b37_phiX174_short/genome/index.html
@@ -172,6 +172,41 @@ Default HG19 reference genome used by GPF
         <td class="nowrap">302.0 B</td>
         <td class="nowrap">bd5e747b25ab3eca4dafef208c8c2cbf</td>
     </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/1_statistic.yaml'>statistics/1_statistic.yaml</a>
+        </td>
+        <td class="nowrap">313.0 B</td>
+        <td class="nowrap">498eb0e860af29958bef946235b0e1db</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/2_statistic.yaml'>statistics/2_statistic.yaml</a>
+        </td>
+        <td class="nowrap">315.0 B</td>
+        <td class="nowrap">2d914d63f47928c9e3f3784001f860b8</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/X_statistic.yaml'>statistics/X_statistic.yaml</a>
+        </td>
+        <td class="nowrap">305.0 B</td>
+        <td class="nowrap">5fec382575553584edd947ce291dc90d</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/reference_genome_statistic.yaml'>statistics/reference_genome_statistic.yaml</a>
+        </td>
+        <td class="nowrap">596.0 B</td>
+        <td class="nowrap">d72e505b975e82e071437de993156b08</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">54.0 B</td>
+        <td class="nowrap">ba67778c5b7dd39d68fe120945cf8ab4</td>
+    </tr>
 </tbody>
 </table>
