+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/MPC/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=MPC
     width="200"></td>
 
-<td>NO RANGE</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -132,6 +132,27 @@ h3,h4 {
         </td>
         <td class="nowrap">8.46 KB</td>
         <td class="nowrap">558273e259dee3322accc97e92e9d7d1</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_MPC.png'>statistics/histogram_MPC.png</a>
+        </td>
+        <td class="nowrap">10.01 KB</td>
+        <td class="nowrap">c5bccdbfe3dc20fc06699476cfd3e5ca</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_MPC.yaml'>statistics/histogram_MPC.yaml</a>
+        </td>
+        <td class="nowrap">1.84 KB</td>
+        <td class="nowrap">fa34cc11495d17d5e7b82b1e4de087ab</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">4.93 KB</td>
+        <td class="nowrap">d93c8841de74d808bf750125fe134b6b</td>
     </tr>
 </tbody>
 </table>
