+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/internal/scores1/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=score0
     width="200"></td>
 
-<td>(865582.000, 195966607.000)</td>
+<td>(nan, nan)</td>
 
 </tr><tr class="score-definition">
 
@@ -94,7 +94,7 @@ h3,h4 {
     title=score2
     width="200"></td>
 
-<td>(8655.820, 1959666.070)</td>
+<td>(nan, nan)</td>
 
 </tr><tr class="score-definition">
 
@@ -116,7 +116,7 @@ h3,h4 {
     title=score4
     width="200"></td>
 
-<td>(86.558, 19596.661)</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -169,6 +169,76 @@ h3,h4 {
         </td>
         <td class="nowrap">1.32 KB</td>
         <td class="nowrap">26aa9750f0a1ad0458bcc99fe7d71ac2</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score0.png'>statistics/histogram_score0.png</a>
+        </td>
+        <td class="nowrap">10.55 KB</td>
+        <td class="nowrap">141a292e722188d15dd0433620edf930</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score0.yaml'>statistics/histogram_score0.yaml</a>
+        </td>
+        <td class="nowrap">1.99 KB</td>
+        <td class="nowrap">081e29b36eed06481ece65f7f3f74d7f</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score2.png'>statistics/histogram_score2.png</a>
+        </td>
+        <td class="nowrap">10.49 KB</td>
+        <td class="nowrap">328478e8566f1c9e12bd16e566b97e6c</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score2.yaml'>statistics/histogram_score2.yaml</a>
+        </td>
+        <td class="nowrap">2.24 KB</td>
+        <td class="nowrap">7ff5adbe4dde53712fc8ed9dde9cffe6</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score4.png'>statistics/histogram_score4.png</a>
+        </td>
+        <td class="nowrap">10.62 KB</td>
+        <td class="nowrap">fa41bbd14d8c1c143c32c77175e6312e</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score4.yaml'>statistics/histogram_score4.yaml</a>
+        </td>
+        <td class="nowrap">2.3 KB</td>
+        <td class="nowrap">d237d79da81e91f24a671e1c7dabb730</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_score0.yaml'>statistics/min_max_score0.yaml</a>
+        </td>
+        <td class="nowrap">48.0 B</td>
+        <td class="nowrap">c1f3dfe8347110a45dc7344c0bca38da</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_score2.yaml'>statistics/min_max_score2.yaml</a>
+        </td>
+        <td class="nowrap">46.0 B</td>
+        <td class="nowrap">be2dd6bfd508cc0f1e0f68a0ccd54db7</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_score4.yaml'>statistics/min_max_score4.yaml</a>
+        </td>
+        <td class="nowrap">46.0 B</td>
+        <td class="nowrap">928031108cbe0fa41f4ca78f36217139</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">13.63 KB</td>
+        <td class="nowrap">4fa7f590cc6b9cdd992b332ce9b6aae3</td>
     </tr>
 </tbody>
 </table>
