+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/internal/scores_incomplete_coverage/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=score0_incomplete_cov
     width="200"></td>
 
-<td>(31776803.000, 140391002.000)</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -125,6 +125,34 @@ h3,h4 {
         </td>
         <td class="nowrap">987.0 B</td>
         <td class="nowrap">f77946424457cdc1e16ae23ba2062b1d</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score0_incomplete_cov.png'>statistics/histogram_score0_incomplete_cov.png</a>
+        </td>
+        <td class="nowrap">16.73 KB</td>
+        <td class="nowrap">c37eb1c557ec79e8be894d86a9211f7d</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_score0_incomplete_cov.yaml'>statistics/histogram_score0_incomplete_cov.yaml</a>
+        </td>
+        <td class="nowrap">2.09 KB</td>
+        <td class="nowrap">9a19dc320d21fef34dd78fa290bbb13e</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_score0_incomplete_cov.yaml'>statistics/min_max_score0_incomplete_cov.yaml</a>
+        </td>
+        <td class="nowrap">65.0 B</td>
+        <td class="nowrap">eddfe133c23241d1566c939701e16c9e</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">4.89 KB</td>
+        <td class="nowrap">eee157b2f6d2a6736eaf5d78ff804e60</td>
     </tr>
 </tbody>
 </table>
