+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/internal/scores_incomplete_coverage/statistics/stats_hash
@@ -9,8 +9,228 @@
     ],
     "histograms": [
       {
-        "type": "null",
-        "reason": "Invalid histogram type"
+        "config": {
+          "type": "number",
+          "view_range": {
+            "min": 31776803.0,
+            "max": 140391002.0
+          },
+          "number_of_bins": 100,
+          "x_log_scale": false,
+          "y_log_scale": false,
+          "x_min_log": null
+        },
+        "bins": [
+          31776803.0,
+          32862944.99,
+          33949086.98,
+          35035228.97,
+          36121370.96,
+          37207512.95,
+          38293654.94,
+          39379796.93,
+          40465938.92,
+          41552080.91,
+          42638222.9,
+          43724364.89,
+          44810506.879999995,
+          45896648.87,
+          46982790.86,
+          48068932.85,
+          49155074.84,
+          50241216.83,
+          51327358.82,
+          52413500.81,
+          53499642.8,
+          54585784.79,
+          55671926.78,
+          56758068.769999996,
+          57844210.76,
+          58930352.75,
+          60016494.739999995,
+          61102636.730000004,
+          62188778.72,
+          63274920.71,
+          64361062.7,
+          65447204.69,
+          66533346.68,
+          67619488.67,
+          68705630.66,
+          69791772.65,
+          70877914.64,
+          71964056.63,
+          73050198.62,
+          74136340.61,
+          75222482.6,
+          76308624.59,
+          77394766.58,
+          78480908.57,
+          79567050.56,
+          80653192.55,
+          81739334.53999999,
+          82825476.53,
+          83911618.52,
+          84997760.50999999,
+          86083902.5,
+          87170044.49000001,
+          88256186.47999999,
+          89342328.47,
+          90428470.46000001,
+          91514612.45,
+          92600754.44,
+          93686896.43,
+          94773038.42,
+          95859180.41,
+          96945322.4,
+          98031464.39,
+          99117606.38,
+          100203748.37,
+          101289890.36,
+          102376032.35,
+          103462174.34,
+          104548316.33,
+          105634458.32,
+          106720600.31,
+          107806742.3,
+          108892884.29,
+          109979026.28,
+          111065168.27,
+          112151310.26,
+          113237452.25,
+          114323594.24,
+          115409736.23,
+          116495878.22,
+          117582020.21,
+          118668162.2,
+          119754304.19,
+          120840446.17999999,
+          121926588.17,
+          123012730.16,
+          124098872.15,
+          125185014.14,
+          126271156.13,
+          127357298.12,
+          128443440.11,
+          129529582.1,
+          130615724.09,
+          131701866.08,
+          132788008.07,
+          133874150.06,
+          134960292.05,
+          136046434.04,
+          137132576.03,
+          138218718.01999998,
+          139304860.01,
+          140391002.0
+        ],
+        "bars": [
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1,
+          0,
+          0,
+          1,
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          2,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1,
+          0,
+          0,
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1
+        ],
+        "out_of_range_bins": [
+          0,
+          0
+        ],
+        "min_value": NaN,
+        "max_value": NaN
       }
     ],
     "table": {
