+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg19/phastCons100/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=phastCons100
     width="200"></td>
 
-<td>NO RANGE</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -111,6 +111,27 @@ h3,h4 {
         </td>
         <td class="nowrap">622.0 B</td>
         <td class="nowrap">c2c4b109900b9f5ea940563bab3a8ad7</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_phastCons100.png'>statistics/histogram_phastCons100.png</a>
+        </td>
+        <td class="nowrap">14.07 KB</td>
+        <td class="nowrap">efcdd4628dbf6987139798f1c30e21e3</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_phastCons100.yaml'>statistics/histogram_phastCons100.yaml</a>
+        </td>
+        <td class="nowrap">403.0 B</td>
+        <td class="nowrap">47c4c84f0598f19fd3ca842c3f2f28c9</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">1.78 KB</td>
+        <td class="nowrap">bc8c83e9f2c3b16a8db0306b442ea368</td>
     </tr>
 </tbody>
 </table>
