+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/GRCh38-hg38/gene_models/refSeq_20200330/index.html
@@ -96,6 +96,20 @@ Default gene models used by GPF for HG38.
         <td class="nowrap">29.84 KB</td>
         <td class="nowrap">05d8bbd9616407e862d82593d5fd1756</td>
     </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats.yaml'>statistics/stats.yaml</a>
+        </td>
+        <td class="nowrap">185.0 B</td>
+        <td class="nowrap">088f8feb9e5264dc7579fe0556a891c7</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">128.0 B</td>
+        <td class="nowrap">fca08b9d5b481f89f0f04ea303d5bea7</td>
+    </tr>
 </tbody>
 </table>
