+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/GRCh38-hg38/genome/index.html
@@ -169,6 +169,34 @@ Default HG38 reference genome used by GPF
         <td class="nowrap">339.0 B</td>
         <td class="nowrap">b151ecdcef68d9014ab78823b5147711</td>
     </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/chr1_statistic.yaml'>statistics/chr1_statistic.yaml</a>
+        </td>
+        <td class="nowrap">324.0 B</td>
+        <td class="nowrap">65e05af202b257ce85a2c3194f4179a0</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/chr2_statistic.yaml'>statistics/chr2_statistic.yaml</a>
+        </td>
+        <td class="nowrap">326.0 B</td>
+        <td class="nowrap">479a740a7f18f059abe9768c651aa272</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/reference_genome_statistic.yaml'>statistics/reference_genome_statistic.yaml</a>
+        </td>
+        <td class="nowrap">578.0 B</td>
+        <td class="nowrap">381346727633b156c5297f852d11b4e5</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">54.0 B</td>
+        <td class="nowrap">160e5784ac6c6e266ab261b64c4ae02a</td>
+    </tr>
 </tbody>
 </table>
