+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/TESTCADD/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=cadd_raw
     width="200"></td>
 
-<td>(0.363, 0.558)</td>
+<td>(nan, nan)</td>
 
 </tr><tr class="score-definition">
 
@@ -94,7 +94,7 @@ h3,h4 {
     title=cadd_phred
     width="200"></td>
 
-<td>(6.202, 8.178)</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -133,6 +133,55 @@ h3,h4 {
         </td>
         <td class="nowrap">601.0 B</td>
         <td class="nowrap">6ddfd7783ed4abd33d3e443bfc751350</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_phred.png'>statistics/histogram_cadd_phred.png</a>
+        </td>
+        <td class="nowrap">14.12 KB</td>
+        <td class="nowrap">23329d335421d56e3eead7b1c6cbd900</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_phred.yaml'>statistics/histogram_cadd_phred.yaml</a>
+        </td>
+        <td class="nowrap">1.99 KB</td>
+        <td class="nowrap">5b2743a6ac2d8d0998dcd0c4a6c3cafa</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_raw.png'>statistics/histogram_cadd_raw.png</a>
+        </td>
+        <td class="nowrap">13.96 KB</td>
+        <td class="nowrap">61c0f45f4a8354e1a9f368222591ba90</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_cadd_raw.yaml'>statistics/histogram_cadd_raw.yaml</a>
+        </td>
+        <td class="nowrap">2.08 KB</td>
+        <td class="nowrap">b0aa57e0a591b8080233002f97bcc14e</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_cadd_phred.yaml'>statistics/min_max_cadd_phred.yaml</a>
+        </td>
+        <td class="nowrap">43.0 B</td>
+        <td class="nowrap">6f2a0cc14be29b8b4303f43697ea206b</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_cadd_raw.yaml'>statistics/min_max_cadd_raw.yaml</a>
+        </td>
+        <td class="nowrap">46.0 B</td>
+        <td class="nowrap">598ee4948b026d0dc780d3fa169540f5</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">9.09 KB</td>
+        <td class="nowrap">7421f8c2eb63ef9a060fc0ef6158db0d</td>
     </tr>
 </tbody>
 </table>
