+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/TESTFreq/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=altFreq
     width="200"></td>
 
-<td>(0.100, 9.423)</td>
+<td>(nan, nan)</td>
 
 </tr><tr class="score-definition">
 
@@ -94,7 +94,7 @@ h3,h4 {
     title=altFreq2
     width="200"></td>
 
-<td>(0.800, 10.435)</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -133,6 +133,55 @@ h3,h4 {
         </td>
         <td class="nowrap">544.0 B</td>
         <td class="nowrap">5057dfbead3499d29983468e36735f88</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_altFreq.png'>statistics/histogram_altFreq.png</a>
+        </td>
+        <td class="nowrap">10.3 KB</td>
+        <td class="nowrap">a20bb25abfcdff63f22fbb2087f540d2</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_altFreq.yaml'>statistics/histogram_altFreq.yaml</a>
+        </td>
+        <td class="nowrap">2.08 KB</td>
+        <td class="nowrap">38c423062aabb20004fc987761fd4122</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_altFreq2.png'>statistics/histogram_altFreq2.png</a>
+        </td>
+        <td class="nowrap">10.64 KB</td>
+        <td class="nowrap">99a701ae7ae9134111deee6c9ffbbff1</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_altFreq2.yaml'>statistics/histogram_altFreq2.yaml</a>
+        </td>
+        <td class="nowrap">2.03 KB</td>
+        <td class="nowrap">43e5e902c11a722578b1898960200f61</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_altFreq.yaml'>statistics/min_max_altFreq.yaml</a>
+        </td>
+        <td class="nowrap">39.0 B</td>
+        <td class="nowrap">2a42a37fb04a15244e3648c1a19e4100</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_altFreq2.yaml'>statistics/min_max_altFreq2.yaml</a>
+        </td>
+        <td class="nowrap">41.0 B</td>
+        <td class="nowrap">052893976a56a5f8ecc5b595da4763fc</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">9.15 KB</td>
+        <td class="nowrap">9910e22bbf9ebaacdfebf887b11f8a18</td>
     </tr>
 </tbody>
 </table>
