+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/TESTPosAgg/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=test_score
     width="200"></td>
 
-<td>(0.000, 0.010)</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -111,6 +111,34 @@ h3,h4 {
         </td>
         <td class="nowrap">99.0 B</td>
         <td class="nowrap">313f112e74853fa39dc6af0fa526b692</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_test_score.png'>statistics/histogram_test_score.png</a>
+        </td>
+        <td class="nowrap">11.97 KB</td>
+        <td class="nowrap">305f3f4472bd42d08a1c26e493505048</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_test_score.yaml'>statistics/histogram_test_score.yaml</a>
+        </td>
+        <td class="nowrap">1.95 KB</td>
+        <td class="nowrap">6ec9ed0142fc0905644864d65c0a69d2</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/min_max_test_score.yaml'>statistics/min_max_test_score.yaml</a>
+        </td>
+        <td class="nowrap">40.0 B</td>
+        <td class="nowrap">e51e615c1d3d01c7c9a49126bce20fb4</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">4.73 KB</td>
+        <td class="nowrap">b2cb34f34f7c74d57b62d45e24fe4aa0</td>
     </tr>
 </tbody>
 </table>
