+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/TESTPosAgg/statistics/stats_hash
@@ -11,8 +11,228 @@
     ],
     "histograms": [
       {
-        "type": "null",
-        "reason": "Invalid histogram type"
+        "config": {
+          "type": "number",
+          "view_range": {
+            "min": 0.0,
+            "max": 0.01
+          },
+          "number_of_bins": 100,
+          "x_log_scale": false,
+          "y_log_scale": false,
+          "x_min_log": null
+        },
+        "bins": [
+          0.0,
+          0.0001,
+          0.0002,
+          0.00030000000000000003,
+          0.0004,
+          0.0005,
+          0.0006000000000000001,
+          0.0007,
+          0.0008,
+          0.0009000000000000001,
+          0.001,
+          0.0011,
+          0.0012000000000000001,
+          0.0013000000000000002,
+          0.0014,
+          0.0015,
+          0.0016,
+          0.0017000000000000001,
+          0.0018000000000000002,
+          0.0019,
+          0.002,
+          0.0021000000000000003,
+          0.0022,
+          0.0023,
+          0.0024000000000000002,
+          0.0025,
+          0.0026000000000000003,
+          0.0027,
+          0.0028,
+          0.0029000000000000002,
+          0.003,
+          0.0031000000000000003,
+          0.0032,
+          0.0033,
+          0.0034000000000000002,
+          0.0035,
+          0.0036000000000000003,
+          0.0037,
+          0.0038,
+          0.0039000000000000003,
+          0.004,
+          0.0041,
+          0.004200000000000001,
+          0.0043,
+          0.0044,
+          0.0045000000000000005,
+          0.0046,
+          0.0047,
+          0.0048000000000000004,
+          0.0049,
+          0.005,
+          0.0051,
+          0.005200000000000001,
+          0.0053,
+          0.0054,
+          0.0055000000000000005,
+          0.0056,
+          0.0057,
+          0.0058000000000000005,
+          0.0059,
+          0.006,
+          0.0061,
+          0.006200000000000001,
+          0.0063,
+          0.0064,
+          0.006500000000000001,
+          0.0066,
+          0.0067,
+          0.0068000000000000005,
+          0.006900000000000001,
+          0.007,
+          0.0071,
+          0.007200000000000001,
+          0.0073,
+          0.0074,
+          0.007500000000000001,
+          0.0076,
+          0.0077,
+          0.0078000000000000005,
+          0.0079,
+          0.008,
+          0.0081,
+          0.0082,
+          0.0083,
+          0.008400000000000001,
+          0.0085,
+          0.0086,
+          0.008700000000000001,
+          0.0088,
+          0.0089,
+          0.009000000000000001,
+          0.0091,
+          0.0092,
+          0.009300000000000001,
+          0.0094,
+          0.0095,
+          0.009600000000000001,
+          0.0097,
+          0.0098,
+          0.0099,
+          0.01
+        ],
+        "bars": [
+          2,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          3,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          1
+        ],
+        "out_of_range_bins": [
+          0,
+          0
+        ],
+        "min_value": NaN,
+        "max_value": NaN
       }
     ],
     "table": {
