+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/TESTphastCons100way/index.html
@@ -72,7 +72,7 @@ h3,h4 {
     title=phastCons100way
     width="200"></td>
 
-<td>NO RANGE</td>
+<td>(nan, nan)</td>
 
 </tr>
 
@@ -139,6 +139,27 @@ h3,h4 {
         </td>
         <td class="nowrap">583.0 B</td>
         <td class="nowrap">56feb878ed78769d5200702392b8e4b7</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_phastCons100way.png'>statistics/histogram_phastCons100way.png</a>
+        </td>
+        <td class="nowrap">11.59 KB</td>
+        <td class="nowrap">7253f59314b3cdec0fe36f4d19a79410</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/histogram_phastCons100way.yaml'>statistics/histogram_phastCons100way.yaml</a>
+        </td>
+        <td class="nowrap">369.0 B</td>
+        <td class="nowrap">b8e089e48e3ecd61146c73b181311056</td>
+    </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">1.81 KB</td>
+        <td class="nowrap">4a80cc48d86c0b4490d3763240c9b871</td>
     </tr>
 </tbody>
 </table>
