+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/hg38/hg38tohg19/index.html
@@ -75,6 +75,13 @@ hg38ToHg19.over.chain.gz
         <td class="nowrap">1.19 MB</td>
         <td class="nowrap">ff3031d93792f4cbb86af44055efd903</td>
     </tr>
+    <tr>
+        <td class="nowrap">
+            <a href='statistics/stats_hash'>statistics/stats_hash</a>
+        </td>
+        <td class="nowrap">11.0 B</td>
+        <td class="nowrap">6a99c575ab87f8c7d1ed1e52e7e349ce</td>
+    </tr>
 </tbody>
 </table>
