+++ dae_conftests/dae_conftests/tests/fixtures/genomic_resources/index.html
@@ -44,8 +44,8 @@
                     <a href='hg19/CADD/index.html'>hg19/CADD</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">3</td>
-                <td class="nowrap">17.61 MB</td>
+                <td class="nowrap">10</td>
+                <td class="nowrap">17.66 MB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -54,8 +54,8 @@
                     <a href='hg19/GATK_ResourceBundle_5777_b37_phiX174_short/gene_models/refGene_201309/index.html'>hg19/GATK_ResourceBundle_5777_b37_phiX174_short/gene_models/refGene_201309</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">2</td>
-                <td class="nowrap">3.86 KB</td>
+                <td class="nowrap">4</td>
+                <td class="nowrap">4.18 KB</td>
                 <td class="nowrap">## refGene gene models from 2013-09
 
 Old gene models used by GPF up to 2019-02.
@@ -67,8 +67,8 @@ Old gene models used by GPF up to 2019-02.
                     <a href='hg19/GATK_ResourceBundle_5777_b37_phiX174_short/genome/index.html'>hg19/GATK_ResourceBundle_5777_b37_phiX174_short/genome</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">3</td>
-                <td class="nowrap">894.05 KB</td>
+                <td class="nowrap">8</td>
+                <td class="nowrap">895.6 KB</td>
                 <td class="nowrap">## HG19 Reference Genome Fragment
 
 Default HG19 reference genome used by GPF
@@ -80,8 +80,8 @@ Default HG19 reference genome used by GPF
                     <a href='hg19/MPC/index.html'>hg19/MPC</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">6</td>
-                <td class="nowrap">418.45 KB</td>
+                <td class="nowrap">9</td>
+                <td class="nowrap">435.23 KB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -90,8 +90,8 @@ Default HG19 reference genome used by GPF
                     <a href='hg19/internal/scores1/index.html'>hg19/internal/scores1</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">5</td>
-                <td class="nowrap">2.93 KB</td>
+                <td class="nowrap">15</td>
+                <td class="nowrap">54.89 KB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -100,8 +100,8 @@ Default HG19 reference genome used by GPF
                     <a href='hg19/internal/scores_incomplete_coverage/index.html'>hg19/internal/scores_incomplete_coverage</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">5</td>
-                <td class="nowrap">2.21 KB</td>
+                <td class="nowrap">9</td>
+                <td class="nowrap">25.98 KB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -110,8 +110,8 @@ Default HG19 reference genome used by GPF
                     <a href='hg19/phastCons100/index.html'>hg19/phastCons100</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">3</td>
-                <td class="nowrap">2.44 MB</td>
+                <td class="nowrap">6</td>
+                <td class="nowrap">2.45 MB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -120,8 +120,8 @@ Default HG19 reference genome used by GPF
                     <a href='hg38/GRCh38-hg38/gene_models/refSeq_20200330/index.html'>hg38/GRCh38-hg38/gene_models/refSeq_20200330</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">2</td>
-                <td class="nowrap">30.04 KB</td>
+                <td class="nowrap">4</td>
+                <td class="nowrap">30.34 KB</td>
                 <td class="nowrap">## refSeq gene models for HG38 from 2020-03
 
 Default gene models used by GPF for HG38.
@@ -133,7 +133,7 @@ Default gene models used by GPF for HG38.
                     <a href='hg38/GRCh38-hg38/genome/index.html'>hg38/GRCh38-hg38/genome</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">3</td>
+                <td class="nowrap">7</td>
                 <td class="nowrap">2.71 MB</td>
                 <td class="nowrap">## HG38 Reference Genome 
 
@@ -146,8 +146,8 @@ Default HG38 reference genome used by GPF
                     <a href='hg38/TESTCADD/index.html'>hg38/TESTCADD</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">3</td>
-                <td class="nowrap">1.31 KB</td>
+                <td class="nowrap">10</td>
+                <td class="nowrap">42.62 KB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -156,8 +156,8 @@ Default HG38 reference genome used by GPF
                     <a href='hg38/TESTFreq/index.html'>hg38/TESTFreq</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">3</td>
-                <td class="nowrap">850.0 B</td>
+                <td class="nowrap">10</td>
+                <td class="nowrap">35.1 KB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -166,8 +166,8 @@ Default HG38 reference genome used by GPF
                     <a href='hg38/TESTPosAgg/index.html'>hg38/TESTPosAgg</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">3</td>
-                <td class="nowrap">621.0 B</td>
+                <td class="nowrap">7</td>
+                <td class="nowrap">19.3 KB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -176,8 +176,8 @@ Default HG38 reference genome used by GPF
                     <a href='hg38/TESTphastCons100way/index.html'>hg38/TESTphastCons100way</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">7</td>
-                <td class="nowrap">9.95 KB</td>
+                <td class="nowrap">10</td>
+                <td class="nowrap">23.71 KB</td>
                 <td class="nowrap"></td>
             </tr>
             <tr>
@@ -186,7 +186,7 @@ Default HG38 reference genome used by GPF
                     <a href='hg38/hg38tohg19/index.html'>hg38/hg38tohg19</a>
                 </td>
                 <td class="nowrap">0</td>
-                <td class="nowrap">2</td>
+                <td class="nowrap">3</td>
                 <td class="nowrap">1.19 MB</td>
                 <td class="nowrap"></td>
             </tr>
